requirements.txt
anyio==3.6.1
bcrypt==3.2.2
boto3==1.24.21
botocore==1.27.21
cffi==1.15.1
click==8.1.3
cryptography==37.0.2
ecdsa==0.17.0
fastapi==0.78.0
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
python-multipart==0.0.5
h11==0.13.0
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jose==1.0.0
pyasn1==0.4.8
python-jose==3.3.0
rsa==4.8
PyJWT==1.6.4
MarkupSafe==2.1.1
numpy==1.22.4
pandas==1.4.2
passlib==1.7.4
Pillow==9.1.1
pycparser==2.21
pydantic==1.9.1
python-dateutil==2.8.2
pytz==2022.1
s3transfer==0.6.0
six==1.15.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.18.2
Werkzeug==2.1.2
zipp==3.8.0
keras==2.9.0
tensorflow-cpu==2.9.0

